Moving plans stuff out of the service_instance file
